requirements.txt
google-api-python-client==1.12.8
google-apitools==0.5.31
google-cloud-aiplatform==0.7.1
google-cloud-storage==1.42.3
apache_beam==2.32.0
pandas==1.4.0
tensorflow==2.5.1
